Resolved;  merge issue
