--- jaclang/compiler/passes/main/py_import_pass.py
@@ -1,99 +1,0 @@
-"""Static Import Pass.
-
-This pass statically imports all modules used in import statements in the
-current module. This pass is run before the def/decl pass to ensure that all
-symbols are available for matching.
-"""
-
-import ast as py_ast
-import importlib.util
-import sys
-from os import path
-from typing import Optional
-
-
-import jaclang.compiler.absyntree as ast
-from jaclang.compiler.passes import Pass
-from jaclang.compiler.passes.main import SubNodeTabPass
-from jaclang.settings import settings
-
-
-class PyImportPass(Pass):
-    """Jac statically imports all modules."""
-
-    def before_pass(self) -> None:
-        """Run once before pass."""
-        self.import_table: dict[str, ast.Module] = {}
-
-    def enter_module(self, node: ast.Module) -> None:
-        """Run Importer."""
-        self.cur_node = node
-        self.import_table[node.loc.mod_path] = node
-        self.terminate()  # Turns off auto traversal for deliberate traversal
-        self.run_again = True
-        while self.run_again:
-            self.run_again = False
-            all_imports = self.get_all_sub_nodes(node, ast.ModulePath)
-            for i in all_imports:
-                lang = i.parent_of_type(ast.Import).hint.tag.value
-                if lang == "py" and settings.py_raise:
-                    mod = self.import_py_module(node=i, mod_path=node.loc.mod_path)
-                    if mod:
-                        # self.run_again = True
-                        i.sub_module = mod
-                        i.add_kids_right([mod], pos_update=False)
-                self.enter_module_path(i)
-            SubNodeTabPass(prior=self, input_ir=node)
-        node.mod_deps = self.import_table
-
-    def enter_module_path(self, node: ast.ModulePath) -> None:
-        """Sub objects.
-
-        path: Sequence[Token],
-        alias: Optional[Name],
-        sub_module: Optional[Module] = None,
-        """
-        if node.alias and node.sub_module:
-            node.sub_module.name = node.alias.value
-        # Items matched during def/decl pass
-
-    # Utility functions
-    # -----------------
-
-    def import_py_module(
-        self, node: ast.ModulePath, mod_path: str
-    ) -> Optional[ast.Module]:
-        """Import a module."""
-        from jaclang.compiler.passes.main import PyastBuildPass
-
-        base_dir = path.dirname(mod_path)
-        sys.path.append(base_dir)
-
-        try:
-            # Dynamically import the module
-            spec = importlib.util.find_spec(node.path_str)
-            sys.path.remove(base_dir)
-            if spec and spec.origin and spec.origin not in {None, "built-in", "frozen"}:
-                if spec.origin in self.import_table:
-                    return self.import_table[spec.origin]
-                with open(spec.origin, "r", encoding="utf-8") as f:
-                    # print(f"\nImporting python module {node.path_str}")
-                    mod = PyastBuildPass(
-                        input_ir=ast.PythonModuleAst(
-                            py_ast.parse(f.read()), mod_path=spec.origin
-                        ),
-                    ).ir
-                if mod:
-                    self.import_table[spec.origin] = mod
-                    return mod
-                else:
-                    raise self.ice(
-                        f"Failed to import python module {node.path_str}: {spec.origin}"
-                    )
-        except Exception as e:
-            self.error(
-                f"Failed to import python module {node.path_str}: {e}",
-                node_override=node,
-            )
-            raise e
-        return None
